fix: split failure-resource transitioner test to satisfy backend size limit
Move TestReleaseResourceTokensOnFailure_PreservesConsumedTokenIdentityRegardlessOfInputOrder
into subsystem_transitioner_failure_resource_test.go so subsystem_transitioner_test.go
stays under the 1000-line backend-size gate after story 001 additions.
